Moved Personalised insights display into component
